pyproject.toml
[project]
name = "job-matcher-rag"
version = "0.1.0"
description = "RAG-based job matching system using Azure OpenAI"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.4.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "openai>=1.0.0",
    "streamlit>=1.30.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]